requirements.txt
torch
torchvision
pandas
tqdm
pillow
